pyproject.toml
[project]
name = "interview-system-v2"
version = "0.1.0"
description = "Adaptive interview system for qualitative consumer research"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.26.0",
    "structlog>=24.0.0",
    "pyyaml>=6.0.0",
    "streamlit>=1.31.0",
    "plotly>=5.18.0",
    "networkx>=3.2",
    "python-louvain>=0.16",
    "spacy>=3.7.0",
    "sentence-transformers>=5.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
